use streams, dependent tasks for gulp task ordering
delete remains the same and is listed as a standalone dependency so that
gulp.start can be used afterwards. not a bad tradeoff. The rest of the
gulp tasks dependencies were listed as dependent task (array), and the
src watching that should occur. Gulp.watch used instead of the watch
module, asset building kept separate into tasks.

A build task was also added so that the initial assets can be built
before watching occurs so that there is not a partial-build and
insufficent assets to begin loading the page.